// Frontend/front/src/app/api/model/purchaseDto.ts
/**
 * CubArt API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PurchaseStatusEnum } from './purchaseStatusEnum';


export interface PurchaseDto { 
    id: string;
    name: string;
    supplierId: number;
    supplierName: string;
    productId: number;
    productName: string;
    facilityId: number;
    facilityName: string;
    amount: number;
    quantity: number;
    purchaseStatus: PurchaseStatusEnum;
    dateCreated: string;
}
export namespace PurchaseDto {
}